Relocate MockOptions for reuse in other tests
